Simplify Vulnerability APIs
