--- analyzer/restapi-plugin/src/main/java/eu/fasten/analyzer/restapiplugin/mvn/api/VulnerabilityApiService.java
@@ -1,49 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package eu.fasten.analyzer.restapiplugin.mvn.api;
-import org.springframework.http.ResponseEntity;
-
-import java.util.List;
-
-public interface VulnerabilityApiService {
-    ResponseEntity<String> getAllVulnerabilities(int offset,
-                                                 int limit,
-                                                 List<String> attributes);
-
-    ResponseEntity<String> getVulnerability(String externalId,
-                                            List<String> attributes);
-
-    ResponseEntity<String> getCoordinatesVulnerabilities(List<String> mavenCoordinates,
-                                                         List<String> attributes);
-
-    ResponseEntity<String> getPurls(String externalId,
-                                    int offset,
-                                    int limit);
-
-    ResponseEntity<String> getVulnerableCallables(String externalId,
-                                                  int offset,
-                                                  int limit);
-
-    ResponseEntity<String> getPackageVersionVulnerabilities(String package_name,
-                                                            String package_version,
-                                                            List<String> attributes);
-
-    ResponseEntity<String> getPurlVulnerabilities(String purl,
-                                                  List<String> attributes);
-}
