+++ analyzer/restapi-plugin/src/main/java/eu/fasten/analyzer/restapiplugin/mvn/api/VulnerableCallChainsApi.java
@@ -18,7 +18,16 @@
 
 package eu.fasten.analyzer.restapiplugin.mvn.api;
 
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import eu.fasten.analyzer.restapiplugin.mvn.KnowledgeBaseConnector;
+import eu.fasten.core.data.FastenURI;
+import eu.fasten.core.vulchains.VulnerableCallChain;
+import eu.fasten.core.vulchains.VulnerableCallChainRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Lazy;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -27,22 +36,45 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.Set;
+
 @Lazy
 @RestController
 @RequestMapping("/{forge}/packages/{pkg}/{pkg_ver}/vulnerable-call-chains")
 public class VulnerableCallChainsApi {
 
-    private final VulnerableCallChainsApiService service;
+    private static final Logger logger = LoggerFactory.getLogger(VulnerableCallChainsApi.class);
 
-    public VulnerableCallChainsApi(VulnerableCallChainsApiService service) {
-        this.service = service;
+    private VulnerableCallChainRepository vulnerableCallChainRepository;
+
+    public VulnerableCallChainsApi() {
+        try {
+            vulnerableCallChainRepository = new VulnerableCallChainRepository(KnowledgeBaseConnector.vulnerableCallChainsPath);
+        } catch (Exception e) {
+            logger.error("Error constructing Vulnerability Call Chain Repository", e);
+            System.exit(1);
+        }
+    }
+
+    /**
+     * Helper method to convert set of vulnerable call chains to JSON formatted string response.
+     * @param chains - a set of {@link VulnerableCallChain} objects to be serialized.
+     * @return {@link String} formatted as JSON response.
+     */
+    private String VulnerableCallChainsToJSON(Set<VulnerableCallChain> chains) {
+        GsonBuilder builder = new GsonBuilder();
+        Gson gson = builder.create();
+        return gson.toJson(chains);
     }
 
     @PostMapping(value = "", produces = MediaType.APPLICATION_JSON_VALUE)
     ResponseEntity<String> getChainsForPackage(@PathVariable("forge") String forge,
                                                @PathVariable("pkg") String packageName,
                                                @PathVariable("pkg_ver") String packageVersion) {
-        return service.getChainsForPackage(forge, packageName, packageVersion);
+
+        Set<VulnerableCallChain> chains = vulnerableCallChainRepository.getChainsForPackage(packageName, packageVersion);
+        var result = VulnerableCallChainsToJSON(chains);
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
     @PostMapping(value = "/module", produces = MediaType.APPLICATION_JSON_VALUE)
@@ -50,7 +82,11 @@ public class VulnerableCallChainsApi {
                                                @PathVariable("pkg") String packageName,
                                                @PathVariable("pkg_ver") String packageVersion,
                                                @RequestBody String rawPath) {
-        return service.getChainsForModule(forge, packageName, packageVersion, rawPath);
+
+        FastenURI fastenUri = FastenURI.create(forge, packageName, packageVersion, rawPath);
+        Set<VulnerableCallChain> chains = vulnerableCallChainRepository.getChainsForModule(fastenUri);
+        var result = VulnerableCallChainsToJSON(chains);
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
     @PostMapping(value = "/callable", produces = MediaType.APPLICATION_JSON_VALUE)
@@ -58,7 +94,10 @@ public class VulnerableCallChainsApi {
                                                 @PathVariable("pkg") String packageName,
                                                 @PathVariable("pkg_ver") String packageVersion,
                                                 @RequestBody String rawPath) {
-        return service.getChainsForCallable(forge, packageName, packageVersion, rawPath);
-    }
 
+        FastenURI fastenUri = FastenURI.create(forge, packageName, packageVersion, rawPath);
+        Set<VulnerableCallChain> chains = vulnerableCallChainRepository.getChainsForCallable(fastenUri);
+        var result = VulnerableCallChainsToJSON(chains);
+        return new ResponseEntity<>(result, HttpStatus.OK);
+    }
 }
